requirements.txt
python-nmap
pyfiglet

